# faster prime sieve part 1