--- a/.gitignore
+++ b/.gitignore
@@ -5,3 +5,4 @@ Manifest.toml
 docs/build
 docs/site
 docs/Manifest.toml
+bench/