--- a/Project.toml
+++ b/Project.toml
@@ -15,4 +15,4 @@ test = ["DataStructures", "IntegerMathUtils", "Test"]
 
 [compat]
 IntegerMathUtils = "0.1.1"
-julia = "1.10"
+julia = "1.6"